requirements.txt
# 核心框架
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# 通义千问模型支持
langchain-dashscope>=0.1.0

# 向量数据库和文档处理
langchain-community>=0.2.0
faiss-cpu>=1.7.4
pypdf>=3.0.0

# 文本处理
langchain-text-splitters>=0.2.0

# 其他必要依赖
numpy>=1.24.0
typing-extensions>=4.5.0 
flask

# 多模态功能依赖
Pillow>=9.0.0 